requirements.txt
flask==3.0.0
google-generativeai==0.3.2
Pillow==10.1.0
python-dotenv==1.0.0
paddleocr==2.7.0.3
paddlepaddle==2.6.0
gunicorn==21.2.0
gradio==4.44.0
